# kvserver: Small optimizations for MaybeGossipNodeLivenessRaftMuLocked