--- a/pkg/gossip/gossip.go
+++ b/pkg/gossip/gossip.go
@@ -47,6 +47,7 @@ the system with minimal total hops. The algorithm is as follows:
 package gossip
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"math"
@@ -891,6 +892,33 @@ func (g *Gossip) AddInfoProto(key string, msg protoutil.Message, ttl time.Durati
 	return g.AddInfo(key, bytes, ttl)
 }
 
+// AddInfoIfNotRedundant adds or updates an info object if it isn't already
+// present in the local infoStore with exactly the same value and with this
+// node as the source. Motivated by the node liveness range's desire to only
+// gossip changed entries.
+//
+// Assumes the values have a TTL of 0 (always adding the gossip anew if the
+// stored value wasn't added with a TTL of 0), because if a value had a non-0
+// TTL then we'd have to make some sort of value judgment about whether it's
+// worth re-gossiping yet given how old the existing matching info was.
+func (g *Gossip) AddInfoIfNotRedundant(key string, val []byte) error {
+	g.mu.Lock()
+	defer g.mu.Unlock()
+
+	info := g.mu.is.getInfo(key)
+
+	if info != nil {
+		infoBytes, err := info.Value.GetBytes()
+		if err == nil && bytes.Equal(infoBytes, val) && g.infoOriginatedHere(info) && info.TTLStamp == math.MaxInt64 {
+			// Nothing has changed, so no need to re-gossip.
+			return nil
+		}
+	}
+
+	// Something is different, so we do need to add the provided key/value.
+	return g.addInfoLocked(key, val, 0 /* ttl */)
+}
+
 // AddClusterID is a convenience method for gossipping the cluster ID. There's
 // no TTL - the record lives forever.
 func (g *Gossip) AddClusterID(val uuid.UUID) error {
@@ -983,6 +1011,12 @@ func (g *Gossip) InfoOriginatedHere(key string) bool {
 	g.mu.RLock()
 	defer g.mu.RUnlock()
 	info := g.mu.is.getInfo(key)
+	return g.infoOriginatedHere(info)
+}
+
+// infoOriginatedHere is a simple reusable helper for InfoOriginatedHere that
+// doesn't involve taking the mutex.
+func (g *Gossip) infoOriginatedHere(info *Info) bool {
 	return info != nil && info.NodeID == g.NodeID.Get()
 }