Full description in the two commit messages, but these are basically attempts to reduce the cost both in terms of CPU cycles and gossip mutex contention when gossiping node liveness from the leaseholder, by both:

1. Doing less protobuf serialization+deserialization
2. Taking the gossip mutex many fewer times

It's debatable whether it's a good idea for the gossip mutex to be taken at all while the node liveness leaseholder `Replica`'s mutex is also held since it seems like a priority inversion, but I didn't really want to boil the whole ocean here at once, and this gets much of the benefit already.

I did a lame microbenchmark (377e8565a1ea61d15610f520bc4ed167accf5aaa) of the old code, the code after the first commit, and the code after the second commit and got:

```
BenchmarkOldCode
BenchmarkOldCode-10                        1526    683569 ns/op
BenchmarkIntermediateCode
BenchmarkIntermediateCode-10               2661    380633 ns/op
BenchmarkNewCode
BenchmarkNewCode-10                        3194    377339 ns/op
```

If I tweaked the code to always have to add the new value to gossip rather than treating it as a no-op, I got:

```
BenchmarkOldCode
BenchmarkOldCode-10                         643   1791176 ns/op
BenchmarkIntermediateCode
BenchmarkIntermediateCode-10                868   1321190 ns/op
BenchmarkNewCode
BenchmarkNewCode-10                         904   1278004 ns/op
```

So it's not exactly world-changing in terms of CPU usage, but (1) it is an improvement, and (2) I'm at least as concerned about possible gossip mutex contention here as I am about cpu usage, and this takes the gossip mutex drastically fewer times than before -- once vs three times when updating a single liveness entry, and once vs (approximately) the number of decommissioned+dead+live nodes times 2 in the bulk case.

Epic: None
(although it marginally helps with #97966)